Fix segmentation fault bug caused by CheatClientAuthentication.
Previously CheatClientAuthentication removed the entries for application_name
from the list of GUC options, to hide the value of application_name.
Which caused a segmentation fault, for example when pg_basebackup was taken.

To fix that bug, this commit makes CheatClientAuthentication reset
application_name to an empty explicitly rather than removing the GUC option
entries for application_name.