Add pg_cheat_funcs.exit_on_segv parameter.
If off, which is the default, segmentation fault will lead to the server crash.
If on, only the current session causing segmentation fault will be terminated.
This is useful to prevent unexpected segmentation fault from causing
the server to crash, for example, when using fragile extensions that are
likely to cause segmentation fault.

This commit also adds pg_segmentation_fault function,
which causes segmentation fault error.