Fix compilation failure caused only with PostgreSQL 15dev.
In PostgreSQL 15dev, the global variable ThisTimeLineID was removed,
and which caused the compilation of pg_xlogfile_name() code to fail
because it used the variable.

To fix the compilation failure, this commit changes pg_xlogfile_name()
so that it uses the current WAL insertion timeline ID during normal
processing and the currently-replaying one during recovery.