Add pg_xid_to_xid8 function.
This function converts the specified transaction ID with 32-bit type xid to
the transaction ID with 64-bit type xid8. The specified transaction ID must
be from the same epoch as the latest transaction ID, or the epoch before.
For example, the currently used xid like pg_stat_activity.backend_xid
or pg_locks.transactionid can be specified in this function.